requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic_core==2.41.5
sentence-transformers==2.2.2
faiss-cpu==1.12.0
google-generativeai==0.3.2
python-multipart==0.0.6
requests==2.31.0
python-dotenv==1.2.1